chore(cleanup): refactor transaction imports and create transaction repository
