+++ src/features/import_mt940/mt940import.py
@@ -9,9 +9,10 @@ from features.account import (account_history_repository
                               as account_history_repository)
 from features.counterparty import (counterparty_repository
                                    as counterparty_repository)
-from features.transaction import (transaction_typ_utils
+from features.transaction import (transaction_typ_repository
                                   as transaction_typ_repository)
-from features.transaction import transaction_utils as transaction_repository
+from features.transaction import (transaction_repository
+                                  as transaction_repository)
 from shared.date_utils import get_iso_date
